atlascope_django/requirements.txt
Django>=5.2.4
django-ninja>=1.4.3
django-cors-headers>=4.7.0
girder-client>=3.2.8
pandas>=2.3.1
